some fixes and merge with main branch
