+++ models/slide.js
@@ -1,0 +1,51 @@
+'use strict';
+const {
+  Model
+} = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+  class Slider extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+    static associate(models) {
+      // define association here
+      Slider.belongsTo(models.Organization, {
+        foreignKey: 'organizationId',
+        as: 'organizations'
+      })
+    }
+  };
+  Slider.init({
+    imageUrl: {
+      type: DataTypes.STRING,
+      allowNull: false,
+      unique: true,
+    },
+    text: {
+      type: DataTypes.TEXT,
+      allowNull: false,
+    },
+    order: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+    organizationId: {
+      type: DataTypes.INTEGER,
+    },
+    deletedAt: {
+      type: DataTypes.DATE,
+    },
+  },
+  {
+    sequelize,
+    modelName: 'Slide',
+    createdAt: 'createdAt',
+    updatedAt: 'updatedAt',
+    deletedAt: 'deletedAt',
+    paranoid: true,
+    timestamps: true,
+  });
+  return Slider;
+};
